refactor(cli): extract manifest planning domain
